requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
playwright==1.49.1
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
python-dotenv==1.0.1
requests==2.32.3
selectolax==0.3.27
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
